==> Proyecto/Versiones Previas/HastaPowerTransform/03_features.ipynb <==
#import pandas as pd
import polars as pl
from datetime import datetime
import warnings

import json

warnings.filterwarnings('ignore')

fase = '03_features'

with open('gen_config.json', 'r') as file:
    gen_config =json.load(file)

folder = gen_config['folder']

#entradas
path_norm = gen_config['path_norm']
path_prod_data = gen_config['path_prod_data']
path_dtw = gen_config['path_dtw']
path_overall_prod_stats = gen_config['path_overall_prod_stats']
#salidas
path_lag = gen_config['path_lag']

#variables config
var_lags = gen_config['var_lags']
var_cates_feat = gen_config['var_cates_feat']

print(f"{'COMIENZA':-^100}")
print(f"{fase:-^100}")

df = pl.read_parquet(f"{folder}/{path_norm}")
prod_data = pl.read_parquet(f"{folder}/{path_prod_data}")
overall_prod_stats = pl.read_parquet(f"{folder}/{path_overall_prod_stats}")
prod_a_predecir = pl.read_csv('..\productos_a_predecir.txt', separator='\t')
df_dtw = pl.read_parquet(f"{folder}/{path_dtw}")

## Pendientes
---
### operaciones sobre tn
* HECHO - razones con lags
* HECHO - marcar cero real
* HECHO - date features:
    - HECHO - mes
    - HECHO - quarter
    - HECHO - mes en el quarter
* HECHO - tiempos de monotonia de la funcion
* maximo salto
* expanding mean

* HECHO - basadas en producto
* HECHO - info del producto
* HECHO porcentual de venta
    - HECHO del mes
    - HECHO del quarter
    - HECHO del año
* porcentual de categoria
* HECHO - stats de categoria1 y 2
* HECHO - stats de brand

## Generales

df = df.with_columns([
    ((pl.col("periodo").cast(pl.Utf8) + "01").str.to_date("%Y%m%d")).alias("periodo_dt"),  
])

df = df.with_columns([
    pl.col('periodo_dt').dt.month().alias('mes'),
    pl.col('periodo_dt').dt.year().alias('year'),
    pl.col('periodo_dt').dt.quarter().alias('quarter'),
    ((pl.col('periodo_dt').dt.month() -1)% 3+1).alias('month_in_quarter'),
])
df = df.with_columns([
    (pl.col("year").cast(pl.Utf8) + "-Q" + pl.col("quarter").cast(pl.Utf8)).alias("yearquarter"),
])
df = df.sort(by=['product_id', 'customer_id', 'periodo_dt'])

periodos = overall_prod_stats.select(['product_id', 'primer_periodo', 'ultimo_periodo'])
df = df.join(periodos, on=['product_id'], how='left', coalesce=True, suffix='_overall')

df = df.with_columns([
    (pl.col('periodo_dt') - pl.col('primer_periodo_overall')).dt.days().alias('edad_dias'),
])

conditional_expr = (
    pl.when((pl.col('primer_periodo_overall') == pl.datetime(2017, 1, 1)))
    .then(pl.lit("Adulto"))
    .when((pl.col('edad_dias') < 35))
    .then(pl.lit("Infante"))
    .when((pl.col('edad_dias') < 190))
    .then(pl.lit("Joven"))
    .when((pl.col('edad_dias') >= 190))
    .then(pl.lit("Adulto"))
    .otherwise(pl.lit("Otro"))
)

df = df.with_columns(conditional_expr.cast(pl.Categorical).alias('edad'))

check = df.select(['periodo_dt', 'primer_periodo_overall', 'edad_dias', 'edad'])
check

# df.filter((pl.col('product_id')==20001) & (pl.col('customer_id')==10001))['tn'].sum()

# df.filter(pl.col('product_id')==20001)['tn'].sum()

## Operaciones sobre tn

# marcar ceros
df = df.with_columns([
    (pl.col('tn') < pl.col('median_tn') / 10 ).alias('tn_cero')
])

lags = var_lags

df = df.sort(by=["product_id", "customer_id", "periodo"])
df = df.set_sorted(["product_id", "customer_id", "periodo"])

start_time = datetime.now()

# lags normales
start_time2 = datetime.now()
for lag in lags:
    df = df.with_columns([
        pl.col("tn_norm").shift(lag).over(["product_id", "customer_id"]).alias(f'tn_lag_{lag}'),
    ])
print(f"Lags creados en: {(datetime.now() - start_time2).total_seconds():.2f} a las {datetime.now().strftime('%H:%M:%S')}", end='\n')

# primer derivada / delta
start_time2 = datetime.now()
for lag in lags:
    df = df.with_columns([
        pl.col("tn_norm").diff(lag).over(["product_id", "customer_id"]).alias(f'tn_diff_{lag}')
    ])
print(f"Diff creados en: {(datetime.now() - start_time2).total_seconds():.2f} a las {datetime.now().strftime('%H:%M:%S')}", end='\n')

# segunda derivada
start_time2 = datetime.now()
for lag in [1,2,3]:
    df = df.with_columns([
        pl.col(f'tn_diff_{lag}').diff(lag).over(["product_id", "customer_id"]).alias(f'tn_diff_2_{lag}')
    ])
print(f"Diff2 creados en: {(datetime.now() - start_time2).total_seconds():.2f} a las {datetime.now().strftime('%H:%M:%S')}", end='\n')

# ratios
start_time2 = datetime.now()
for lag in lags:
    df = df.with_columns([
        ((-pl.col(f'tn_diff_{lag}'))/pl.col(f'tn_lag_{lag}')).alias(f'ratio_{lag}')
    ])
print(f"Ratios creados en: {(datetime.now() - start_time2).total_seconds():.2f} a las {datetime.now().strftime('%H:%M:%S')}", end='\n')

# Promedios
start_time2 = datetime.now()
for lag in lags:
    df = df.with_columns([
        (pl.col('tn_norm').rolling_mean(lag).over(["product_id", "customer_id"])).alias(f'avg_{lag}')
    ])
print(f"Promedios creados en: {(datetime.now() - start_time2).total_seconds():.2f} a las {datetime.now().strftime('%H:%M:%S')}", end='\n')

# Maximos
start_time2 = datetime.now()
for lag in lags:
    df = df.with_columns([
        (pl.col('tn_norm').rolling_max(lag).over(["product_id", "customer_id"]) == pl.col('tn_norm')).alias(f'max_{lag}')
    ])
print(f"Maximos creados en: {(datetime.now() - start_time2).total_seconds():.2f} a las {datetime.now().strftime('%H:%M:%S')}", end='\n')

# Minimos
start_time2 = datetime.now()
for lag in lags:
    df = df.with_columns([
        (pl.col('tn_norm').rolling_min(lag).over(["product_id", "customer_id"]) == pl.col('tn_norm')).alias(f'max_{lag}')
    ])
print(f"Minimos creados en: {(datetime.now() - start_time2).total_seconds():.2f} a las {datetime.now().strftime('%H:%M:%S')}", end='\n')

# Monotonia_creciente
start_time2 = datetime.now()
for lag in range(1,12):
    df = df.with_columns([
        (((pl.col(f'tn_lag_{lag}'))-pl.col(f'tn_lag_{lag+1}')) > 0).alias(f'crece_{lag+1}')
    ])
print(f"Monotonia_creciente creados en: {(datetime.now() - start_time2).total_seconds():.2f} a las {datetime.now().strftime('%H:%M:%S')}", end='\n')

# Monotonia_decreciente
start_time2 = datetime.now()
for lag in range(1,12):
    df = df.with_columns([
        (((pl.col(f'tn_lag_{lag}'))-pl.col(f'tn_lag_{lag+1}')) > 0).alias(f'decrece_{lag+1}')
    ])
print(f"Monotonia_decreciente creados en: {(datetime.now() - start_time2).total_seconds():.2f} a las {datetime.now().strftime('%H:%M:%S')}", end='\n')

print(f"\nFin de creacion de metricas {(datetime.now() - start_time).total_seconds():.2f} a las {datetime.now().strftime('%H:%M:%S')}", end='\n')

## nivel producto

df = df.join(prod_data, on=['product_id'], how='left', coalesce=True, suffix='')

catcols = [col for col in df.columns if df[col].dtype == pl.Utf8]

df = df.with_columns([
    pl.col(col).cast(pl.Categorical) for col in catcols
])

print(f"Columnas cambiadas a categorical: {catcols}")

var_stats = []

for variable in var_cates_feat:
    var_stats = df.groupby('product_id',variable).agg([
        pl.sum('tn').alias(f'{variable}_total'),
        pl.median('tn').alias(f'{variable}_median'),
        pl.max('tn').alias(f'{variable}_max'),
    ])
    df = df.join(var_stats, on=['product_id', variable], how='left', coalesce=True, suffix='')

tiempos = ['periodo','yearquarter','year']

for tiempo in tiempos:
    tiempo_sums = df.group_by([tiempo,'product_id']).agg([pl.sum('tn').alias(f'tn_total_{tiempo}')])
    df = df.join(tiempo_sums, on=[tiempo,'product_id'], how='left', coalesce=True, suffix='')
    df = df.with_columns([
        (pl.col(f'tn') / pl.col(f'tn_total_{tiempo}')).alias(f'prop_product_{tiempo}')
    ])
    df = df.drop(columns=f'tn_total_{tiempo}')

for tiempo in tiempos:
    for variable in var_cates_feat:
        combi_sums = df.group_by([tiempo,'product_id', variable]).agg([pl.sum('tn').alias(f'tn_total_{tiempo}_{variable}')])
        df = df.join(combi_sums, on=[tiempo,'product_id', variable], how='left', coalesce=True, suffix='')
        df = df.with_columns([
            (pl.col(f'tn') / pl.col(f'tn_total_{tiempo}_{variable}')).alias(f'prop_product_{tiempo}_{variable}')
        ])
        df = df.drop(columns=f'tn_total_{tiempo}_{variable}')

# props = df.columns
# props = [prop for prop in props if prop.startswith("prop_")]

# start_time = datetime.now()

# # lags normales
# start_time2 = datetime.now()



# print(f"Lags creados en: {(datetime.now() - start_time2).total_seconds():.2f} a las {datetime.now().strftime('%H:%M:%S')}", end='\n')

# print(f"\nFin de creacion de metricas {(datetime.now() - start_time).total_seconds():.2f} a las {datetime.now().strftime('%H:%M:%S')}", end='\n')

# DTW

df = df.join(df_dtw, on=['product_id', 'customer_id'], how='left', coalesce=True, suffix='')

## ajustes finales pre export

df = df.sort(by=["product_id", "customer_id", "periodo"])
df = df.set_sorted(["product_id", "customer_id", "periodo"])

#df.to_parquet('sell_in_lag.parquet', index=False)
df.write_parquet(f'{folder}/{path_lag}')

print(f"{fase:-^100}")
print(f"{'FINALIZA':-^100}\n\n\n")
