==> Proyecto/Versiones Previas/HastaPowerTransform/04_crear_train_set.ipynb <==
#import pandas as pd
import polars as pl
from datetime import datetime
import json

import warnings
warnings.filterwarnings('ignore')

fase = '04_crear_train_test'

with open('gen_config.json', 'r') as file:
    gen_config =json.load(file)

folder = gen_config['folder']

#entradas
path_lag = gen_config['path_lag']
path_overall_prod_stats = gen_config['path_overall_prod_stats']
#salidas
path_train = gen_config['path_train']
path_test = gen_config['path_test']
path_futuro = gen_config['path_futuro']

#variables config
periodo_train_start = gen_config['var_periodo_train_start']
periodo_train_end = gen_config['var_periodo_train_end']
periodo_test = gen_config['var_periodo_test']
periodo_futuro = gen_config['var_periodo_futuro']

print(f"{'COMIENZA':-^100}")
print(f"{fase:-^100}")

overall_prod_stats = pl.read_parquet(f"{folder}/{path_overall_prod_stats}")
overall_prod_stats = overall_prod_stats[['product_id','tot_weight', 'avg_weight', 'med_weight']]
df = pl.read_parquet(f"{folder}/{path_lag}")
df = df.join(overall_prod_stats, on=['product_id'], how='left')

print(f"Tamaño datos con features: {df.shape}")

df = df.with_columns([
    pl.col('primer_periodo').cast(pl.Int32),
    pl.col('ultimo_periodo').cast(pl.Int32),
    pl.col('periodo_dt').cast(pl.Int32),
    (-pl.col("tn_norm").diff(-2).over(["product_id", "customer_id"])).alias(f'tn_futuro')
])


df_train = df.filter((pl.col('periodo') >= periodo_train_start) & (pl.col('periodo') <= periodo_train_end))
df_test = df.filter(pl.col('periodo') == periodo_test)
df_futuro = df.filter(pl.col('periodo') == periodo_futuro)

print(f"{'Shape Train':.<25}: {df_train.shape} para los periodos desde {periodo_train_start} hasta {periodo_train_end}")
print(f"{'Shape Test':.<25}: {df_test.shape} para el periodo {periodo_test}")
print(f"{'Shape Futuro':.<25}: {df_futuro.shape} para el periodo {periodo_futuro}")

df_train.write_parquet(f'{folder}/{path_train}')
df_test.write_parquet(f'{folder}/{path_test}')
df_futuro.write_parquet(f'{folder}/{path_futuro}')

df_train.glimpse()

print(f"{fase:-^100}")
print(f"{'FINALIZA':-^100}\n\n\n")
