==> Proyecto/Versiones Previas/HastaPowerTransform/07_semillerio.ipynb <==
import pandas as pd
import os
import json

fase = '07_semillerio'

with open('gen_config.json', 'r') as file:
    gen_config =json.load(file)

folder = gen_config['folder']

path_semillerio = gen_config['path_semillerio']

print(f"{'COMIENZA':-^100}")
print(f"{fase:-^100}")

# csv1 = pd.read_csv('./Current Best/kaggle.csv')

# csv2 = pd.read_csv('./Current Best/kaggle-new.csv')

# combined_df = pd.merge(csv1, csv2, on='product_id')

# # Mostrar el DataFrame combinado para verificar
# print("DataFrame combinado:")
# print(combined_df.head())


# import polars as pl 

# # Convertir el DataFrame de Pandas a Polars
# df = pl.DataFrame(combined_df)
 
# # Definir las columnas a promediar
# columnas_a_promediar = ['tn_x', 'tn_y']
 
# # Calcular el promedio sumando las columnas y dividiendo por la cantidad de columnas
# df = df.with_columns(
#     (sum(pl.col(col) for col in columnas_a_promediar) / len(columnas_a_promediar)).alias('promedio_tn')
# )
 
# df= df.select(['product_id', 'promedio_tn'])

# df

import glob

# Ruta a la carpeta actual
carpeta = './Current Best/'

# Obtener una lista de todos los archivos CSV en la carpeta actual
archivos_csv = glob.glob(os.path.join(carpeta, '*.csv'))

print(archivos_csv)

dfs = []
# Iterar sobre cada archivo en la lista y leerlo en un DataFrame
for archivo in archivos_csv:
    df = pd.read_csv(archivo)
    dfs.append(df)

dfs2 = pd.concat(dfs, ignore_index=True)
dfs_agrupados = dfs2.groupby('product_id')['tn'].mean()

dfs

df

type(dfs_agrupados)

dfs_agrupados=pd.DataFrame(dfs_agrupados, columns=['tn'])

dfs_agrupados

dfs_agrupados=dfs_agrupados.reset_index()

dfs_agrupados

dfs_agrupados.to_csv('semillerio.csv', index=False)

dfs_agrupados.to_csv(f"{folder}/{datetime.now().strftime('%Y%m%d_%H%M%S')}_{path_kaggle}", index=False)

print(f"{fase:-^100}")
print(f"{'FINALIZA':-^100}\n\n\n")
